# src/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: Shopix E-Commerce API
  version: "1.0.0"
  description: |
    Full REST API documentation for the **Shopix** e-commerce platform.

    ### Authentication
    Most endpoints require a valid `Accesstoken` cookie set after login.
    Pass credentials via the login endpoint — the server sets HttpOnly cookies automatically.

  contact:
    name: Shopix Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: Accesstoken
  schemas:
...
    Order:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        storeId:
          type: string
        productId:
          type: string
        receiverName:
          type: string
        mobileNumber:
          type: string
        billingAddress:
          type: string
        quantity:
          type: integer
        totalAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, stripe]
        paymentStatus:
          type: string
          enum: [Pending, Paid]
        status:
          type: string
          enum: [new, processing, shipped, completed, cancelled]
        returnStatus:
          type: string
          enum: [none, processing, successful, failed]
        createdAt:
          type: string
          format: date-time

    # ─── Notification ─────────────────────────────────────────
...
paths:
...
  /api/orders/return:
    post:
      tags: [Returns]
      summary: Submit a return request
      description: Buyer submits a return request for a completed order. Optional photos can be uploaded.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [orderId, reason]
              properties:
                orderId:
                  type: string
                reason:
                  type: string
                  example: Product was damaged on arrival
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Return request submitted
        "400":
          description: Invalid order state, or return already submitted
        "403":
          description: Unauthorized — not the buyer

    get:
      tags: [Returns]
      summary: Get return requests for a store (seller)
      description: Returns all orders with a return status for the given store.
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: storeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Return orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"

    patch:
      tags: [Returns]
      summary: Accept or decline return (seller)
      description: Seller accepts or rejects a buyer's return request. Notifies buyer via email and in-app notification.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, decision]
              properties:
                orderId:
                  type: string
                decision:
                  type: string
                  enum: [successful, failed]
      responses:
        "200":
          description: Return decision recorded
        "400":
          description: Missing fields
        "404":
          description: Order not found

  # ───────────────────────────────────────────────────────────
  # REVIEWS
  # ───────────────────────────────────────────────────────────
...